# src/server/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Twilio Account Proxy Service",
    "description": "A proxy service to create and manage accounts and subaccounts on the Twilio platform.",
    "version": "1.0.0"
  },
  "paths": {
    "/account/{accountSid}/subAccounts.json": {
      "get": {
        "description": "Returns all subaccounts pertaining to account provided",
        "responses": {
          "200": {
            "description": "A list of subaccounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "uniqueItems": true,
                  "items": {
                    "$ref": "#/components/schemas/Subaccount"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "accountSid",
            "in": "path",
            "required": true,
            "description": "A parent account which may have an arbitraty number of subaccounts.",
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ]
      }
    },
    "/account/{accountSid}/subAccounts/createSubaccount.json": {
      "post": {
        "description": "Returns a new subaccount under the account provided",
        "responses": {
          "201": {
            "description": "Successfully created a new subaccount",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subaccount"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "accountSid",
            "in": "path",
            "description": "A parent account which may have an arbitraty number of subaccounts.",
            "schema": {
              "type": "string"
            },
            "style": "simple",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubaccount"
              }
            }
          }
        }
      }
    },
    "/account/{accountSid}/subaccounts/{subaccountSid}.json": {
      "get": {
        "description": "Returns the subaccount corresponding to the accountSid and subaccountSid provided",
        "responses": {
          "200": {
            "description": "Successfully returned subaccount",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subaccount"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "accountSid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          },
          {
            "name": "subaccountSid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Subaccount": {
        "type": "object",
        "properties": {
          "sid": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "friendlyName": {
            "type": "string"
          },
          "parentAccountSid": {
            "type": "string"
          },
          "dateTimeCreated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateSubaccount": {
        "type": "object",
        "properties": {
          "friendlyName": {
            "type": "string"
          },
          "accountSid": {
            "type": "string"
          }
        },
        "required": [
          "friendlyName",
          "accountSid"
        ]
      }
    }
  }
}